+++ .github/workflows/ReleaseNotes.md
@@ -1,6 +1,7 @@
 ## What's Changed?
 
 * Changed RGB to save as RGB10 instead of RGB255
+* Added context menu with "duplicate" option to make a duplicate of moves and palettes
 
 ## Known Bugs